Combine inReplyToRoot into MessageToUser.  We'll ignore the fact that comments will eventually need something similar/exactly the same.
