--- lib/jdomodels/src/main/resources/schema/MessageInReplyToRoot-ddl.sql
@@ -1,8 +1,0 @@
-CREATE TABLE `MESSAGE_IN_REPLY_TO_ROOT` (
-  `ROOT_MESSAGE_ID` bigint(20) NOT NULL,
-  `CHILD_MESSAGE_ID` bigint(20) NOT NULL,
-  PRIMARY KEY (`ROOT_MESSAGE_ID`, `CHILD_MESSAGE_ID`),
-  UNIQUE KEY (`CHILD_MESSAGE_ID`), 
-  CONSTRAINT `MESSAGE_ROOT_SELF_FK` FOREIGN KEY (`ROOT_MESSAGE_ID`) REFERENCES `MESSAGE_IN_REPLY_TO_ROOT` (`CHILD_MESSAGE_ID`) ON DELETE CASCADE,
-  CONSTRAINT `MESSAGE_ROOT_ID_FK` FOREIGN KEY (`CHILD_MESSAGE_ID`) REFERENCES `MESSAGE_CONTENT` (`ID`) ON DELETE CASCADE
-)
