+++ lib/jdomodels/src/main/resources/schema/MessageToUser-ddl.sql
@@ -1,6 +1,8 @@
 CREATE TABLE `MESSAGE_TO_USER` (
   `MESSAGE_ID` bigint(20) NOT NULL,
   `SUBJECT` varchar(256) DEFAULT NULL, 
+  `ROOT_MESSAGE_ID` bigint(20) NOT NULL,
   PRIMARY KEY (`MESSAGE_ID`),
-  CONSTRAINT `MESSAGE_TO_USER_FK` FOREIGN KEY (`MESSAGE_ID`) REFERENCES `MESSAGE_CONTENT` (`ID`) ON DELETE CASCADE
+  CONSTRAINT `MESSAGE_TO_USER_FK` FOREIGN KEY (`MESSAGE_ID`) REFERENCES `MESSAGE_CONTENT` (`ID`) ON DELETE CASCADE, 
+  CONSTRAINT `MESSAGE_TO_USER_SELF_FK` FOREIGN KEY (`ROOT_MESSAGE_ID`) REFERENCES `MESSAGE_TO_USER` (`MESSAGE_ID`) ON DELETE CASCADE
 )
