+++ lib/lib-auto-generated/src/main/resources/schema/org/sagebionetworks/repo/model/migration/MigrationType.json
@@ -7,10 +7,11 @@
     	"CREDENTIAL", 
     	"USER_PROFILE",
         "FILE_HANDLE",
-        "MESSAGE", 
-        "MESSAGE_THREAD", 
-        "MESSAGE_THREAD_OBJECT", 
+        "MESSAGE_CONTENT", 
+        "MESSAGE_TO_USER", 
+        "MESSAGE_RECIPIENT", 
         "MESSAGE_STATUS", 
+        "COMMENT", 
         "WIKI_PAGE",
         "WIKI_ATTACHMENT",
         "WIKI_OWNERS",
